requirements: separate gateware build from runtime operation.
Effectively splits requirements for two use-cases:
* building Entangler gateware
* running the Entangler software from an ARTIQ experiment.
	The experiment doesn't need migen to run, it just needs to know
	the PHY addresses which are solely dependent on Entangler settings.
