+++ entangler/phy.py
@@ -1,10 +1,8 @@
 """Gateware-side ARTIQ RTIO interface to the entangler core."""
-import enum
 import logging
 import math
 import typing
 
-import numpy
 from artiq.gateware.rtio import rtlink
 from migen import Case
 from migen import Cat
@@ -16,38 +14,9 @@ from migen import Signal
 
 from entangler.core import EntanglerCore
 from entangler.config import settings
+from entangler.phy_registers import ADDRESS_WRITE, max_value_to_bit_width
 
 _LOGGER = logging.getLogger(__name__)
-
-# generate the PHY read/write addresses, b/c ARTIQ kernel had issues w/ referencing dynaconf settings
-def max_value_to_bit_width(max_value: int) -> int:
-    """Calculate how many bits are needed to represent an unsigned int."""
-    return math.ceil(math.log2(max_value))
-
-
-_num_channels = settings.NUM_ENTANGLER_INPUT_SIGNALS + settings.NUM_OUTPUT_CHANNELS
-_channel_bits = max_value_to_bit_width(_num_channels)
-_read_start = 0b1 << (_channel_bits + 1)
-
-
-class ADDRESS_WRITE(enum.IntEnum):
-    """PHY Addresses to configure the Entangler."""
-
-    CONFIG = 0
-    RUN = 1
-    TCYCLE = 2
-    PATTERNS = 3
-    TIMING = 0b1 << _channel_bits
-
-
-class ADDRESS_READ(enum.IntEnum):
-    """PHY Addresses to get information from the Entangler."""
-
-    STATUS = _read_start + 0
-    NCYCLES = _read_start + 1
-    TIME_REMAINING = _read_start + 2
-    NTRIGGERS = _read_start + 3
-    TIMESTAMP = numpy.int32(0b11 << _channel_bits)
 
 
 class Entangler(Module):
