+++ nix/default.nix
@@ -1,5 +1,6 @@
 { pkgs ? import <nixpkgs> {}
 , artiqpkgs ? import <artiq-full> {}
+, buildGateware ? true  # whether the requirements for building gateware should be added. Setting to false reduces requirements, but is useful if you just want to use the coredevice drivers in experiments.
 }:
 
 # To run this package in development mode (where code changes are reflected in shell w/o restart),
@@ -25,18 +26,20 @@ in
     propagatedBuildInputs = [
       artiqpkgs.artiq
       entangler-deps.dynaconf
-      python3Packages.jsonschema
+    ] ++ lib.optionals buildGateware [
+      python3Packages.jsonschema # really a dependency of artiq.coredevice.jsondesc, but that's imported by entangler.kasli_generic so it's functionally a requirement
       python3Packages.mergedeep
       artiqpkgs.migen
       artiqpkgs.misoc
     ];
 
-    doCheck = true;
+    doCheck = buildGateware;  # all pytest tests are actually gateware tests, require migen
     checkInputs = [ python3Packages.pytestCheckHook ];
     pytestFlagsArray = [
       "-m 'not slow'"
     ];
-    pythonImportsCheck = [ pname "${pname}.kasli_generic" "${pname}.driver" "${pname}.phy" ];
+    pythonImportsCheck = [ pname "${pname}.driver" ]
+      ++ lib.optionals buildGateware [ "${pname}.kasli_generic" "${pname}.core" "${pname}.phy" ];
 
     meta = with lib; {
       description = "ARTIQ extension to generate & check patterns (for entanglement).";
