+++ setup.py
@@ -10,18 +10,18 @@ if __name__ == "__main__":
         name="entangler",
         version="1.1.1",
         packages=setuptools.find_packages(),
-        requirements=[
-            "artiq>=5",
-            "dynaconf>=3",
-            "mergedeep",
-            "migen",
-            "misoc",
-            "numpy",
-        ],
+        requirements=["artiq>=5", "dynaconf>=3", "mergedeep", "numpy"],
         url="https://github.com/drewrisinger/entangler-core",
         setup_requires=["pytest-runner"],
         tests_require=["pytest"],
         package_data={
             "": ["*.txt", "*.toml", "*.md", "*.json"],
+        },
+        extras_require={
+            "gateware": [
+                "jsonschema",
+                "migen",
+                "misoc",
+            ]
         },
     )
